Update psquare.py
I think this was a small bug in the initialisation of psquare. The algorithm should not update positions and heights when it inititalises. 

Example bad case:
```
values = [-0.93440463,  1.762478  , -0.13665276, -0.50844987,  0.69265]
psquare = PSquare(50)
for val in values:
    psquare.update(val)
psquare.marker_positions
```
gives us `array([1, 2, 3, 4, 6])` which is not right (as max marker position becomes greater than the the number of values entered)